support ascend using infer_ext
