+++ requirements/runtime.txt
@@ -17,5 +17,9 @@ tiktoken
 torch<=2.3.1,>=2.0.0
 torchvision<=0.18.1,>=0.15.0
 transformers
+<<<<<<< HEAD
 triton>=2.1.0,<=2.3.1; sys_platform == "linux"
+=======
+triton>=2.1.0,<=2.2.0; sys_platform == "linux"
+>>>>>>> support ascend using infer_ext
 uvicorn
